Add support for the Nucleo STM32L432KC Board.
This needed quite a bit of re-work in the PLL / RCC configuration as the
STM32Lxx device differs a lot from the STM32F4xx devices supported so far.

Also, the new code makes use of a lot of compile time checks so it needs
C++14 or maybe even C++17.